pyproject.toml
[tool.poetry]
name = "ticket_system"
version = "0.1.0"
description = "A railway ticket reservation system"
authors = ["Prince Kumar <neo11prince@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
django = "^3.2"
djangorestframework = "^3.12"
psycopg2 = "^2.9"
environ = "^1.0"
django-environ = "^0.12.0"
drf-yasg = "^1.21.9"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
gunicorn = "^20.1"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"

[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''
preview = true

[tool.isort]
profile = "black"
skip = ["migrations"]
multi_line_output = 0
force_single_line = false
ensure_newline_before_comments = true
line_length = 120
combine_as_imports = true
include_trailing_comma = false
force_grid_wrap = 0
use_parentheses = false
lines_between_types = 1
group_by_package = true
from_first = true
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
no_lines_before = ["LOCALFOLDER"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
